NP-3279 Extract the part that builds the Reference object from the CristiMapper and into its own class.
